rag/requirements.txt
gradio
PyPDF2
sentence-transformers
faiss-cpu
transformers
groq
